requirements.txt
PyMuPDF==1.23.26
fpdf2==2.7.4